functions/requirements.txt
firebase_functions~=0.1.0
google-cloud-logging
firebase-admin
stripe==12.5.1
python-dotenv
pydantic==2.11.9
